requirements.txt
PyYAML
torch
tensorflow
tensorboard
tensorboardX
